+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/model/validator/M_HU.java
@@ -16,7 +16,7 @@ import de.metas.handlingunits.IHandlingUnitsBL;
 import de.metas.handlingunits.IHandlingUnitsDAO;
 import de.metas.handlingunits.exceptions.HUException;
 import de.metas.handlingunits.model.I_M_HU;
-import de.metas.inoutcandidate.api.IShipmentScheduleInvalidateBL;
+import de.metas.inoutcandidate.invalidation.IShipmentScheduleInvalidateBL;
 import de.metas.logging.LogManager;
 import de.metas.storage.spi.hu.impl.StorageSegmentFromHU;
 import de.metas.util.Services;
