+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/model/validator/M_HU_Storage.java
@@ -27,7 +27,7 @@ import org.adempiere.ad.modelvalidator.annotations.Validator;
 import org.compiere.model.ModelValidator;
 
 import de.metas.handlingunits.model.I_M_HU_Storage;
-import de.metas.inoutcandidate.api.IShipmentScheduleInvalidateBL;
+import de.metas.inoutcandidate.invalidation.IShipmentScheduleInvalidateBL;
 import de.metas.storage.IStorageSegment;
 import de.metas.storage.spi.hu.impl.StorageSegmentFromHUStorage;
 import de.metas.util.Services;
